requirements.txt
flask
openai
python-docx
python-dotenv
requests
PyGithub
jira